Basic tests are working from Unity tests

+++ linker/Tests2/Mono.Linker.Tests.Core/Utils/NiceIO.cs
@@ -4,11 +4,11 @@ using System.IO;
 using System.Linq;
 using System.Text;
 
-namespace NiceIO
+namespace Mono.Linker.Tests.Core.Utils
 {
-    public class NPath : IEquatable<NPath>
+    public class NPath : IEquatable<NPath>, IComparable
     {
-        private static readonly StringComparison PathStringComparison = IsLinux() ? StringComparison.InvariantCulture : StringComparison.InvariantCultureIgnoreCase;
+        private static readonly StringComparison PathStringComparison = IsLinux() ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase;
 
         private readonly string[] _elements;
         private readonly bool _isRelative;
@@ -23,11 +23,26 @@ namespace NiceIO
 
             path = ParseDriveLetter(path, out _driveLetter);
 
-            var split = path.Split('/', '\\');
+            if (path == "/")
+            {
+                _isRelative = false;
+                _elements = new string[] { };
+            }
+            else
+            {
+                var split = path.Split('/', '\\');
 
-            _isRelative = _driveLetter == null && IsRelativeFromSplitString(split);
+                _isRelative = _driveLetter == null && IsRelativeFromSplitString(split);
 
-            _elements = ParseSplitStringIntoElements(split.Where(s => s.Length > 0).ToArray());
+                _elements = ParseSplitStringIntoElements(split.Where(s => s.Length > 0).ToArray());
+            }
+        }
+
+        private NPath(string[] elements, bool isRelative, string driveLetter)
+        {
+            _elements = elements;
+            _isRelative = isRelative;
+            _driveLetter = driveLetter;
         }
 
         private string[] ParseSplitStringIntoElements(IEnumerable<string> inputs)
@@ -36,7 +51,12 @@ namespace NiceIO
 
             foreach (var input in inputs.Where(input => input.Length != 0))
             {
-                if (input == "..")
+                if (input == ".")
+                {
+                    if ((stack.Count > 0) && (stack.Last() != "."))
+                        continue;
+                }
+                else if (input == "..")
                 {
                     if (HasNonDotDotLastElement(stack))
                     {
@@ -76,13 +96,6 @@ namespace NiceIO
             return split[0].Length != 0 || !split.Any(s => s.Length > 0);
         }
 
-        private NPath(string[] elements, bool isRelative, string driveLetter)
-        {
-            _elements = elements;
-            _isRelative = isRelative;
-            _driveLetter = driveLetter;
-        }
-
         public NPath Combine(params string[] append)
         {
             return Combine(append.Select(a => new NPath(a)).ToArray());
@@ -112,13 +125,36 @@ namespace NiceIO
         public NPath RelativeTo(NPath path)
         {
             if (!IsChildOf(path))
-                throw new ArgumentException("Path.RelativeTo() was invoked with two paths that are unrelated. invoked on: " + ToString() + " asked to be made relative to: " + path);
+            {
+                if (!IsRelative && !path.IsRelative && _driveLetter != path._driveLetter)
+                    throw new ArgumentException("Path.RelativeTo() was invoked with two paths that are on different volumes. invoked on: " + ToString() + " asked to be made relative to: " + path);
+
+                NPath commonParent = null;
+                foreach (var parent in RecursiveParents)
+                {
+                    commonParent = path.RecursiveParents.FirstOrDefault(otherParent => otherParent == parent);
+
+                    if (commonParent != null)
+                        break;
+                }
+
+                if (commonParent == null)
+                    throw new ArgumentException("Path.RelativeTo() was unable to find a common parent between " + ToString() + " and " + path);
+
+                if (IsRelative && path.IsRelative && commonParent.IsEmpty())
+                    throw new ArgumentException("Path.RelativeTo() was invoked with two relative paths that do not share a common parent.  Invoked on: " + ToString() + " asked to be made relative to: " + path);
+
+                var depthDiff = path.Depth - commonParent.Depth;
+                return new NPath(Enumerable.Repeat("..", depthDiff).Concat(_elements.Skip(commonParent.Depth)).ToArray(), true, null);
+            }
 
             return new NPath(_elements.Skip(path._elements.Length).ToArray(), true, null);
         }
 
         public NPath ChangeExtension(string extension)
         {
+            ThrowIfRoot();
+
             var newElements = (string[])_elements.Clone();
             newElements[newElements.Length - 1] = Path.ChangeExtension(_elements[_elements.Length - 1], WithDot(extension));
             if (extension == string.Empty)
@@ -136,7 +172,12 @@ namespace NiceIO
 
         public string FileName
         {
-            get { return _elements.Last(); }
+            get
+            {
+                ThrowIfRoot();
+
+                return _elements.Last();
+            }
         }
 
         public string FileNameWithoutExtension
@@ -147,6 +188,11 @@ namespace NiceIO
         public IEnumerable<string> Elements
         {
             get { return _elements; }
+        }
+
+        public int Depth
+        {
+            get { return _elements.Length; }
         }
 
         public bool Exists(string append = "")
@@ -183,6 +229,9 @@ namespace NiceIO
         {
             get
             {
+                if (IsRoot)
+                    throw new ArgumentException("A root directory does not have an extension");
+
                 var last = _elements.Last();
                 var index = last.LastIndexOf(".");
                 if (index < 0) return String.Empty;
@@ -207,6 +256,10 @@ namespace NiceIO
 
         public string ToString(SlashMode slashMode)
         {
+            // Check if it's linux root /
+            if (IsRoot && string.IsNullOrEmpty(_driveLetter))
+                return Slash(slashMode).ToString();
+
             if (_isRelative && _elements.Length == 0)
                 return ".";
 
@@ -303,6 +356,14 @@ namespace NiceIO
             }
         }
 
+        public int CompareTo(object obj)
+        {
+            if (obj == null)
+                return -1;
+
+            return this.ToString().CompareTo(((NPath)obj).ToString());
+        }
+
         public static bool operator !=(NPath a, NPath b)
         {
             return !(a == b);
@@ -323,6 +384,12 @@ namespace NiceIO
         {
             return _elements.Length == 0;
         }
+
+        public bool IsRoot
+        {
+            get { return _elements.Length == 0 && !_isRelative; }
+        }
+
         #endregion inspection
 
         #region directory enumeration
@@ -363,6 +430,7 @@ namespace NiceIO
         public NPath CreateFile()
         {
             ThrowIfRelative();
+            ThrowIfRoot();
             EnsureParentDirectoryExists();
             File.WriteAllBytes(ToString(), new byte[0]);
             return this;
@@ -383,6 +451,10 @@ namespace NiceIO
         public NPath CreateDirectory()
         {
             ThrowIfRelative();
+
+            if (IsRoot)
+                throw new NotSupportedException("CreateDirectory is not supported on a root level directory because it would be dangerous:" + ToString());
+
             Directory.CreateDirectory(ToString());
             return this;
         }
@@ -466,6 +538,9 @@ namespace NiceIO
         {
             ThrowIfRelative();
 
+            if (IsRoot)
+                throw new NotSupportedException("Delete is not supported on a root level directory because it would be dangerous:" + ToString());
+
             if (FileExists())
                 File.Delete(ToString());
             else if (DirectoryExists())
@@ -482,9 +557,21 @@ namespace NiceIO
                 throw new InvalidOperationException("Trying to delete a path that does not exist: " + ToString());
         }
 
+        public void DeleteIfExists(DeleteMode deleteMode = DeleteMode.Normal)
+        {
+            ThrowIfRelative();
+
+            if (FileExists() || DirectoryExists())
+                Delete(deleteMode);
+        }
+
         public NPath DeleteContents()
         {
             ThrowIfRelative();
+
+            if (IsRoot)
+                throw new NotSupportedException("DeleteContents is not supported on a root level directory because it would be dangerous:" + ToString());
+
             if (FileExists())
                 throw new InvalidOperationException("It is not valid to perform this operation on a file");
 
@@ -526,6 +613,10 @@ namespace NiceIO
         public NPath Move(NPath dest)
         {
             ThrowIfRelative();
+
+            if (IsRoot)
+                throw new NotSupportedException("Move is not supported on a root level directory because it would be dangerous:" + ToString());
+
             if (dest.IsRelative)
                 return Move(Parent.Combine(dest));
 
@@ -566,7 +657,7 @@ namespace NiceIO
             {
                 if (Path.DirectorySeparatorChar == '\\')
                     return new NPath(Environment.GetEnvironmentVariable("USERPROFILE"));
-                return new NPath(Environment.GetFolderPath(Environment.SpecialFolder.Personal));
+                return new NPath(Environment.GetEnvironmentVariable("HOME"));
             }
         }
 
@@ -584,6 +675,12 @@ namespace NiceIO
         {
             if (_isRelative)
                 throw new ArgumentException("You are attempting an operation on a Path that requires an absolute path, but the path is relative");
+        }
+
+        private void ThrowIfRoot()
+        {
+            if (IsRoot)
+                throw new ArgumentException("You are attempting an operation that is not valid on a root level directory");
         }
 
         public NPath EnsureDirectoryExists(string append = "")
@@ -608,6 +705,22 @@ namespace NiceIO
             return parent;
         }
 
+        public NPath FileMustExist()
+        {
+            if (!FileExists())
+                throw new FileNotFoundException("File was expected to exist : " + ToString());
+
+            return this;
+        }
+
+        public NPath DirectoryMustExist()
+        {
+            if (!DirectoryExists())
+                throw new DirectoryNotFoundException("Expected directory to exist : " + ToString());
+
+            return this;
+        }
+
         public bool IsChildOf(string potentialBasePath)
         {
             return IsChildOf(new NPath(potentialBasePath));
@@ -617,6 +730,14 @@ namespace NiceIO
         {
             if ((IsRelative && !potentialBasePath.IsRelative) || !IsRelative && potentialBasePath.IsRelative)
                 throw new ArgumentException("You can only call IsChildOf with two relative paths, or with two absolute paths");
+
+            // If the other path is the root directory, then anything is a child of it as long as it's not a Windows path
+            if (potentialBasePath.IsRoot)
+            {
+                if (_driveLetter != potentialBasePath._driveLetter)
+                    return false;
+                return true;
+            }
 
             if (IsEmpty())
                 return false;
@@ -631,8 +752,6 @@ namespace NiceIO
         {
             get
             {
-                ThrowIfRelative();
-
                 var candidate = this;
                 while (true)
                 {
@@ -652,6 +771,8 @@ namespace NiceIO
 
         public NPath ParentContaining(NPath needle)
         {
+            ThrowIfRelative();
+
             return RecursiveParents.FirstOrDefault(p => p.Exists(needle));
         }
 
@@ -691,6 +812,9 @@ namespace NiceIO
 
         public IEnumerable<NPath> MoveFiles(NPath destination, bool recurse, Func<NPath, bool> fileFilter = null)
         {
+            if (IsRoot)
+                throw new NotSupportedException("MoveFiles is not supported on this directory because it would be dangerous:" + ToString());
+
             destination.EnsureDirectoryExists();
             return Files(recurse).Where(fileFilter ?? AlwaysTrue).Select(file => file.Move(destination.Combine(file.RelativeTo(this)))).ToArray();
         }
